+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ### Bug Fixes
 
-* include download urls in userscript metadata ([#2](https://github.com/popstas/userscripts/issues/2)) ([92a467f](https://github.com/popstas/userscripts/commit/92a467f431f3d3eba513e97d8e245509430657c6))
+* debounce codex-helper single task completion event 10cd9ac
+* include download urls in userscript metadata (#2) 92a467f, closes #2